+++ src/kernels.cu
@@ -49,8 +49,8 @@ __global__ void countNNZPerColumn(const int* col_indices, int* col_counts, int n
 }
 
 // Kernel to scatter values and row indices to transposed matrix
-__global__ void scatterToTransposed(const int* values, const int* col_indices, const int* row_ptr,
-                                    int* t_values, int* t_row_indices, int* t_col_ptr, int num_rows) {
+__global__ void scatterToTransposed(const float* values, const int* col_indices, const int* row_ptr,
+                                    float* t_values, int* t_row_indices, int* t_col_ptr, int num_rows) {
     int row = blockIdx.x * blockDim.x + threadIdx.x;
     if (row < num_rows) {
         for (int j = row_ptr[row]; j < row_ptr[row + 1]; ++j) {
